refactor: extract meta home component
