+++ src/components/pages/home/index.js
@@ -1,4 +1,5 @@
 import Hero from './hero'
 import Screenshots from './screenshots'
+import Meta from './Meta'
 
-export { Hero, Screenshots }
+export { Hero, Screenshots, Meta }
